Move ingress hook to separate file
